Release 7.1

+++ blender/source/blender/blenkernel/intern/colortools.c
@@ -50,7 +50,6 @@
 
 
 #include "IMB_colormanagement.h"
-#include "IMB_imbuf.h"
 #include "IMB_imbuf_types.h"
 
 /* ********************************* color curve ********************* */
@@ -577,14 +576,14 @@ static void curvemap_make_table(CurveMap *cuma, const rctf *clipr)
 	/* store first and last handle for extrapolation, unit length */
 	cuma->ext_in[0] = bezt[0].vec[0][0] - bezt[0].vec[1][0];
 	cuma->ext_in[1] = bezt[0].vec[0][1] - bezt[0].vec[1][1];
-	range = sqrt(cuma->ext_in[0] * cuma->ext_in[0] + cuma->ext_in[1] * cuma->ext_in[1]);
+	range = sqrtf(cuma->ext_in[0] * cuma->ext_in[0] + cuma->ext_in[1] * cuma->ext_in[1]);
 	cuma->ext_in[0] /= range;
 	cuma->ext_in[1] /= range;
 
 	a = cuma->totpoint - 1;
 	cuma->ext_out[0] = bezt[a].vec[1][0] - bezt[a].vec[2][0];
 	cuma->ext_out[1] = bezt[a].vec[1][1] - bezt[a].vec[2][1];
-	range = sqrt(cuma->ext_out[0] * cuma->ext_out[0] + cuma->ext_out[1] * cuma->ext_out[1]);
+	range = sqrtf(cuma->ext_out[0] * cuma->ext_out[0] + cuma->ext_out[1] * cuma->ext_out[1]);
 	cuma->ext_out[0] /= range;
 	cuma->ext_out[1] /= range;
