+++ blender/source/blender/render/intern/source/pointdensity.c
@@ -55,7 +55,6 @@
 #include "DNA_particle_types.h"
 
 #include "render_types.h"
-#include "renderdatabase.h"
 #include "texture.h"
 #include "pointdensity.h"
 
@@ -384,7 +383,7 @@ static void accum_density(void *userdata, int index, float squared_dist)
 	else if (pdr->falloff_type == TEX_PD_FALLOFF_CONSTANT)
 		density = pdr->squared_radius;
 	else if (pdr->falloff_type == TEX_PD_FALLOFF_ROOT)
-		density = sqrt(dist);
+		density = sqrtf(dist);
 	else if (pdr->falloff_type == TEX_PD_FALLOFF_PARTICLE_AGE) {
 		if (pdr->point_data_used & POINT_DATA_LIFE)
 			density = dist*MIN2(pdr->point_data[pdr->offset + index], 1.0f);
